// src/api/model/menuListItemDto.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * PizzaApp | v1
 * OpenAPI spec version: 1.0.0
 */

export interface MenuListItemDto {
  id?: string;
  name: string;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  pizzasCount?: number | string;
}