// src/api/model/menuDetailsDto.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * PizzaApp | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PizzaSearchResultDto } from './pizzaSearchResultDto';

export interface MenuDetailsDto {
  pizzas?: PizzaSearchResultDto[];
  id?: string;
  name: string;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  pizzasCount?: number | string;
}