// src/api/model/createMenuDto.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * PizzaApp | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LookUpItemDto } from './lookUpItemDto';

export interface CreateMenuDto {
  pizzeria: LookUpItemDto;
  /** @maxLength 50 */
  name: string;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
}